requirements.txt
dash==2.10.0
dash_bootstrap_components==1.4.1
garminconnect==0.1.55
numpy==1.21.5
pandas==2.0.2
plotly==5.6.0
PyYAML==6.0